Split RPMVesselComputer into 3 files
Slightly easier to navigate than one 3500 line monsterfile.
